refactor: extract shared frontend components (DRY) (#306)
- Extract ExpandedAssetChart shared component
- Extract useDebouncedValue hook
- Extract useTrackedSymbols hook
- Extract SearchResultItem component
- Use existing formatChangePct in pages
