+++ frontend/src/hooks/use-debounced-value.ts
@@ -1,0 +1,16 @@
+import { useState, useEffect } from "react"
+
+/**
+ * Returns a debounced version of the given value.
+ * The returned value only updates after `delay` ms of inactivity.
+ */
+export function useDebouncedValue<T>(value: T, delay: number): T {
+  const [debounced, setDebounced] = useState(value)
+
+  useEffect(() => {
+    const timer = setTimeout(() => setDebounced(value), delay)
+    return () => clearTimeout(timer)
+  }, [value, delay])
+
+  return debounced
+}
